+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 1.3.0 - 2022-08-22
+### Added
+- Some more tests
+- VHost env to rabbit mq container
+- Method to get container
+### Changed
+- Remove redmine container configuration
+- Change coding style
+- Update README.md
+
 ## 1.2.0 - 2022-07-26
 ### Added
 - Redis container